pyproject.toml
[project]
name = "phidata-recruitment-backend"
version = "0.1.0"
description = "FastAPI backend for AI Recruitment System"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.42.0",
    "mistralai>=1.2.5",
    "openai>=1.58.1",
    "phidata>=2.7.5",
    "pypdf2>=3.0.1",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "redis>=5.0.1",
    "faiss-cpu>=1.7.4",
    "python-dotenv>=1.0.1",
    "pydantic>=2.5.3",
    "numpy>=1.26.3",
    "requests>=2.31.0",
    "python-multipart>=0.0.9"
]
